Fail quietly when not enough electrons to excite
The 20-from-ci.py example doesn't work when there are less than 4 alpha or 4 beta electrons because of this line of code. 

In the case of a system like H4, then the preferred behavior might be to return something that evaluates to False, so then we can set aaaa or bbbb blocks to zero, while still keeping the aabb blocks.