Improve flexibility of mcscf.addons.state_average
The mcscf.addons.state_average function now creates a child class of the
casscf function argument as well as the casscf.fcisolver attribute. This
allows the creation of children classes of state-averaged CASCI and
MCSCF calculation objects. The values of the weights are now attributes
of the children objects and can be addressed after object creation (they
are no longer "lost" in the local scope of the addons.state_average_
function).