Improve flexibility of mcscf.addons.state_average
The mcscf.addons.state_average function now creates a child class of the
casscf function argument as well as the casscf.fcisolver attribute. This
allows the creation of children metaclasses for state-averaged CASCI and
MCSCF calculation objects. The values of the weights are now attributes
of the children objects and can be addressed after object creation (they
are no longer "lost" in the local scope of the addons.state_average_
function). The second-order CASSCF algorithm is extended to state-
averaged CASSCF calculations based on these metaclasses, with a view
towards implementing the Lagrange-multiplier method of SACASSCF
single-root nuclear gradients as described in Mol. Phys. 99, 103 (2001).