Locally alias time functions in logger
This commit changes how the compatibility aliasing is done in the
pyscf logger. To accommodate differing time APIs the module was
previously overriding methods in stdlib time to accommodate changes in
behavior between python 2.x and newer version of python 3.x. However,
changing functions and overriding methods in an external module
(especially a stdlib module) is problematic and can break other
people's code using pyscf depending on documented behavior of the
external libraries. To fix this and not cause unexpected side effects
for users this changes how this compatibility aliasing is done, it uses
a local alias in the module and sets things to call the correct method
depending on the python version.

Fixes #814