df/grad/rhf.py refactor: remove ishf kwarg
"Is this an HF calculation?" is a terrible way to solve a problem.
It's better to redesign so that things happen in their natural
places. The caller now is responsible for tagging the get_jk
density matrix argument with mo_coeff and mo_occ, and for summing
the contributions arising from different density matrices according
to whether it's RHF, UKS, or anything else in the "get_veff"
function.