smearing for ROHF and fix_spin=True
For ROHF and spin>0, PySCF's smearing function produces incorrect results (and drives spin to 0), because it assumes the meaning of the occupation vector as in RHF. I propose a solution where the smearing function is applied separately to alpha and beta occupations which are later added up to produce the ROHF occupation vector. This is obviously wrong for low multiplicities and high smearing sigma, but it's "almost" accurate for high enough multiplicity and low enough sigma.